Refactoring the message cell context menu handling.
