+++ Shared/UI/Views/MessageContentView/MessageContentView.swift
@@ -24,7 +24,9 @@ class MessageContentView: BaseView {
     }
     static let authorViewHeight: CGFloat = 40
 
-    static var standardHeight: CGFloat { return MessageContentView.bubbleHeight - MessageContentView.textViewPadding }
+    static var standardHeight: CGFloat {
+        return MessageContentView.bubbleHeight - MessageContentView.textViewPadding
+    }
     static let padding = Theme.ContentOffset.long
     static var textViewPadding: CGFloat { return MessageContentView.padding.value * 2 }
