+++ iOS/Flows/Conversation/Cells/MessageCell.swift
@@ -18,6 +18,7 @@ struct MessageDetailState: Equatable {
 /// A cell for displaying individual messages, author and reactions.
 class MessageCell: UICollectionViewCell {
 
+    // Input handlers
     var handleTappedMessage: ((ConversationId, MessageId) -> Void)?
     var handleEditMessage: ((ConversationId, MessageId) -> Void)?
     var handleTappedAttachment: ((MediaItem) -> Void)?
@@ -25,7 +26,7 @@ class MessageCell: UICollectionViewCell {
     let content = MessageContentView()
 
     // Detail View
-    @ObservedObject private var messageState = MessageDetailViewState(message: nil)
+    @ObservedObject var messageState = MessageDetailViewState(message: nil)
     private lazy var detailView = MessageDetailView(config: self.messageState)
     private lazy var detailVC = NavBarIgnoringHostingController(rootView: self.detailView)
     var shouldShowDetailBar: Bool = true
@@ -33,6 +34,9 @@ class MessageCell: UICollectionViewCell {
     /// If true, this cell's message details are fully visible.
     @Published private(set) var messageDetailState = MessageDetailState()
     private var subscriptions = Set<AnyCancellable>()
+
+    // Context menu
+    private lazy var contextMenuDelegate = MessageCellContextMenuDelegate(messageCell: self)
 
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -47,7 +51,7 @@ class MessageCell: UICollectionViewCell {
     private func initializeViews() {
         self.contentView.addSubview(self.content)
 
-        let contextMenuInteraction = UIContextMenuInteraction(delegate: self)
+        let contextMenuInteraction = UIContextMenuInteraction(delegate: self.contextMenuDelegate)
         self.content.bubbleView.addInteraction(contextMenuInteraction)
 
         ConversationsManager.shared.$activeConversation
@@ -173,106 +177,6 @@ class MessageCell: UICollectionViewCell {
 
             try? await message.setToConsumed()
         }.add(to: self.messageDetailTasks)
-    }
-}
-
-// MARK: - Context Menu
-
-extension MessageCell: UIContextMenuInteractionDelegate {
-
-    func contextMenuInteraction(_ interaction: UIContextMenuInteraction,
-                                configurationForMenuAtLocation location: CGPoint) -> UIContextMenuConfiguration? {
-
-        return UIContextMenuConfiguration(identifier: nil) { () -> UIViewController? in
-            guard let message = self.messageState.message else { return nil }
-            return MessagePreviewViewController(with: message)
-        } actionProvider: { (suggestions) -> UIMenu? in
-            return self.makeContextMenu()
-        }
-    }
-
-    private func makeContextMenu() -> UIMenu {
-        guard let message = self.messageState.message as? Message, let cid = message.cid else {
-            return UIMenu()
-        }
-
-        let neverMind = UIAction(title: "Never Mind", image: UIImage(systemName: "nosign")) { action in }
-
-        let confirmDelete = UIAction(title: "Confirm",
-                                     image: UIImage(systemName: "trash"),
-                                     attributes: .destructive) { action in
-            Task {
-                let controller = ChatClient.shared.messageController(cid: cid, messageId: message.id)
-                do {
-                    try await controller.deleteMessage()
-                } catch {
-                    logError(error)
-                }
-            }
-        }
-
-        let deleteMenu = UIMenu(title: "Delete Message",
-                                image: UIImage(systemName: "trash"),
-                                options: .destructive,
-                                children: [confirmDelete, neverMind])
-
-        let viewReplies = UIAction(title: "View Replies") { [unowned self] action in
-            self.handleTappedMessage?(cid, message.id)
-        }
-
-        let edit = UIAction(title: "Edit",
-                            image: UIImage(systemName: "pencil.circle")) { [unowned self] action in
-            self.handleEditMessage?(cid, message.id)
-        }
-
-        let read = UIAction(title: "Set to read",
-                            image: UIImage(systemName: "eyeglasses")) { [unowned self] action in
-            self.setToRead()
-        }
-
-        let unread = UIAction(title: "Set to unread",
-                            image: UIImage(systemName: "eyeglasses")) { [unowned self] action in
-            self.setToUnread()
-        }
-
-        var menuElements: [UIMenuElement] = []
-
-        if !isRelease, message.isFromCurrentUser {
-            menuElements.append(deleteMenu)
-        }
-
-        if !isRelease, message.isFromCurrentUser {
-            menuElements.append(edit)
-        }
-
-        if message.isConsumedByMe {
-            menuElements.append(unread)
-        } else if message.canBeConsumed {
-            menuElements.append(read)
-        }
-
-        if message.parentMessageId.isNil {
-            menuElements.append(viewReplies)
-        }
-
-        return UIMenu.init(title: "From: \(message.author.parseUser?.fullName ?? "Unkown")",
-                           image: nil,
-                           identifier: nil,
-                           options: [],
-                           children: menuElements)
-    }
-
-    func contextMenuInteraction(_ interaction: UIContextMenuInteraction,
-                                previewForHighlightingMenuWithConfiguration configuration: UIContextMenuConfiguration) -> UITargetedPreview? {
-
-        let params = UIPreviewParameters()
-        params.backgroundColor = ThemeColor.clear.color
-        params.shadowPath = UIBezierPath(rect: .zero)
-        if let bubble = interaction.view as? SpeechBubbleView, let path = bubble.bubbleLayer.path {
-            params.visiblePath = UIBezierPath(cgPath: path)
-        }
-        let preview = UITargetedPreview(view: interaction.view!, parameters: params)
-        return preview
     }
 }
